move add/update/delete comment to components/comments.jsx
